requirements.txt
black~=23.12.1
PyQt6~=6.6.1
python-dateutil~=2.8.2
pyqtdarktheme~=2.1.0
pyinstaller
